refactor: new service to also have separate db scripts. Add own users/roles, decouple from persistence module
