+++ common/persistence/src/main/java/app/coronawarn/server/common/persistence/repository/TraceTimeIntervalWarningRepository.java
@@ -1,6 +1,7 @@
-package app.coronawarn.server.common.persistence.eventregistration.repository;
+package app.coronawarn.server.common.persistence.repository;
 
-import app.coronawarn.server.common.persistence.eventregistration.domain.TraceTimeIntervalWarning;
+
+import app.coronawarn.server.common.persistence.domain.TraceTimeIntervalWarning;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
