+++ common/persistence/src/main/resources/db/specific/postgresql/V9__createPermissionsForEventRegistration.sql
@@ -1,2 +1,1 @@
-GRANT SELECT, INSERT ON TABLE trace_location TO "cwa_event_registration";
 GRANT INSERT ON TABLE trace_time_interval_warning TO "cwa_submission";
