+++ services/event-registration/src/main/java/app/coronawarn/server/services/eventregistration/boundary/TraceLocationController.java
@@ -6,8 +6,8 @@ import static app.coronawarn.server.services.eventregistration.config.UrlConstan
 import app.coronawarn.server.common.protocols.internal.evreg.TraceLocation;
 import app.coronawarn.server.services.eventregistration.boundary.validation.ValidTraceLocation;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.context.request.async.DeferredResult;
@@ -15,8 +15,8 @@ import org.springframework.web.context.request.async.DeferredResult;
 
 @RestController
 @RequestMapping(V1)
+@Validated
 public class TraceLocationController {
-
 
   /**
    * Entrypoint for creating new trace location.
@@ -26,11 +26,11 @@ public class TraceLocationController {
    */
   @PostMapping(path = TRACE_LOCATION_ROUTE, consumes = "application/x-protobuf")
   public DeferredResult<ResponseEntity<Void>> createTraceLocation(
-      @ValidTraceLocation
-      @RequestBody TraceLocation traceLocation) {
+      @ValidTraceLocation TraceLocation traceLocation) {
     DeferredResult<ResponseEntity<Void>> result = new DeferredResult<>();
     result.setResult(ResponseEntity.noContent().build());
     return result;
   }
+
 
 }
