+++ services/event-registration/src/main/java/app/coronawarn/server/services/eventregistration/domain/TraceLocation.java
@@ -1,4 +1,4 @@
-package app.coronawarn.server.common.persistence.eventregistration.domain;
+package app.coronawarn.server.services.eventregistration.domain;
 
 import org.springframework.data.annotation.Id;
