+++ services/event-registration/src/main/resources/application.yaml
@@ -1,4 +1,24 @@
 ---
+event-registration:
+  version: 0
+
+spring:
+  main:
+    web-application-type: NONE
+  transaction:
+    default-timeout: 20
+  flyway:
+    enabled: true
+    locations: classpath:/eventregistration/db/migration, classpath:/eventregistration/db/specific/{vendor}
+    password: ${POSTGRESQL_PASSWORD_FLYWAY:local_setup_flyway}
+    user: ${POSTGRESQL_USER_FLYWAY:local_setup_flyway}
+  # Postgres configuration
+  datasource:
+    driver-class-name: org.postgresql.Driver
+    url: jdbc:postgresql://${POSTGRESQL_SERVICE_HOST:localhost}:${POSTGRESQL_SERVICE_PORT:5432}/${POSTGRESQL_DATABASE:cwa}
+    username: ${POSTGRESQL_USER_EVENT_REGISTRATION:local_event_registration}
+    password: ${POSTGRESQL_PASSWORD_EVENT_REGISTRATION:local_event_registration}
+
 logging:
   level:
     org:
@@ -11,22 +31,3 @@ logging:
         config:
           sources:
             URLConfigurationSource: ERROR
-event-registration:
-  version: 0
-
-spring:
-  main:
-    web-application-type: NONE
-  transaction:
-    default-timeout: 20
-  flyway:
-    enabled: true
-    locations: classpath:/db/migration, classpath:/db/specific/{vendor}
-    password: ${POSTGRESQL_PASSWORD_FLYWAY:local_setup_flyway}
-    user: ${POSTGRESQL_USER_FLYWAY:local_setup_flyway}
-  # Postgres configuration
-  datasource:
-    driver-class-name: org.postgresql.Driver
-    url: jdbc:postgresql://${POSTGRESQL_SERVICE_HOST:localhost}:${POSTGRESQL_SERVICE_PORT:5432}/${POSTGRESQL_DATABASE:cwa}
-    username: ${POSTGRESQL_USER_UPLOAD:local_event_registration}
-    password: ${POSTGRESQL_PASSWORD_UPLOAD:local_event_registration}
