+++ services/event-registration/src/main/resources/setup/create-users.sql
@@ -1,0 +1,7 @@
+/* Creates CWA users */
+
+CREATE USER "<change me>" WITH INHERIT IN ROLE cwa_flyway ENCRYPTED PASSWORD '<change me>';
+
+
+/* --------------- Event Registration --------------- */
+CREATE USER "<change me>" WITH INHERIT IN ROLE cwa_event_registration ENCRYPTED PASSWORD '<change me>';
