+++ services/event-registration/src/test/java/app/coronawarn/server/services/eventregistration/EventRegistrationApplicationTest.java
@@ -8,5 +8,4 @@ public class EventRegistrationApplicationTest {
   void applicationContextShouldLoad() {
     EventRegistrationApplication.main(new String[]{});
   }
-
 }
