+++ services/event-registration/src/test/java/app/coronawarn/server/services/eventregistration/boundary/TraceLocationControllerTest.java
@@ -1,10 +1,10 @@
 package app.coronawarn.server.services.eventregistration.boundary;
 
 import app.coronawarn.server.common.protocols.internal.evreg.TraceLocation;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 
@@ -14,25 +14,31 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.request;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.springframework.test.web.servlet.setup.MockMvcBuilders.standaloneSetup;
 
-@SpringBootTest
-@AutoConfigureMockMvc
+@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 public class TraceLocationControllerTest {
 
-  @Autowired
+
   private MockMvc mockMvc;
+
+  @BeforeEach
+  void setup() {
+    this.mockMvc = standaloneSetup(TraceLocationController.class).build();
+  }
+
 
   @Test
   public void shouldReturnNoContent() throws Exception {
-    MvcResult mvcResult = this.mockMvc.perform(
+    final TraceLocation build = TraceLocation.newBuilder().build();
+    final MvcResult mvcResult = this.mockMvc.perform(
         post(V1 + TRACE_LOCATION_ROUTE)
             .contentType("application/x-protobuf")
-            .content(TraceLocation.newBuilder().build().toByteArray()))
+            .content(build.toByteArray()))
         .andExpect(request().asyncStarted())
         .andReturn();
 
-    this.mockMvc.perform(asyncDispatch(mvcResult))
-        .andExpect(status().isNoContent());
+    this.mockMvc.perform(asyncDispatch(mvcResult)).andExpect(status().isNoContent());
   }
 
 }
